+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## v50
 ###### *Not yet released*
+* **ST controls**: A simple control panel on the Chronicle dashboard enables STs to end all scenelong effects, recover motes or willpower, and more
+* Mote committments can be marked as scenelong, causing them to lapse when the ST uses the 'end scene' control
+* Better performance in the Character editor
 * Bug fixes (thanks Everspace, Madletter, Artirian_Legacy)
 
 ## v49